Updated Main Menu
I have changed the code so the game will accept all type of inputs from the user, both string and integers to select the options in the Main Menu.

User can type in "1" or "Play/play" to select the play option to play the game.

User can type in "2" or "Quit/quit" to select the quit option to quit the game.